requirement.txt
absl-py==0.8.0
astor==0.8.0
backports.csv==1.0.7
bayesian-optimization==1.0.1
beautifulsoup4==4.8.2
blis==0.4.1
boto==2.49.0
boto3==1.9.243
botocore==1.12.243
certifi==2019.9.11
chardet==3.0.4
cheroot==8.2.1
CherryPy==18.5.0
Click==7.0
cloudpickle==1.2.2
cycler==0.10.0
cymem==2.0.2
docutils==0.15.2
feedparser==5.2.1
Flask==1.1.1
future==0.17.1
futures==3.1.1
gast==0.3.2
gensim==3.8.1
google-pasta==0.1.7
googletrans==2.4.0
goslate==1.5.1
grpcio==1.24.1
gym==0.14.0
h5py==2.10.0
idna==2.8
importlib-resources==1.0.2
itsdangerous==1.1.0
jaraco.classes==3.1.0
jaraco.collections==3.0.0
jaraco.functools==3.0.0
jaraco.text==3.2.0
Jinja2==2.10.3
jmespath==0.9.4
joblib==0.14.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keras-rl2==1.0.3
Kivy==1.11.1
kivy-deps.glew==0.1.12
kivy-deps.gstreamer==0.1.18
kivy-deps.sdl2==0.1.23
Kivy-Garden==0.1.4
kiwisolver==1.1.0
lxml==4.4.1
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mkl-fft==1.0.14
mkl-random==1.1.0
mkl-service==2.3.0
more-itertools==8.1.0
murmurhash==1.0.2
mysqlclient==1.4.6
nltk==3.4.5
numpy==1.17.2
pandas==0.25.1
pandas-datareader==0.8.1
Pattern==3.6
pdfminer.six==20200104
Pillow==6.2.0
Pillow-PIL==0.1.dev0
pipenv==2018.11.26
plac==0.9.6
portend==2.6
preshed==3.0.2
protobuf==3.10.0
pycryptodome==3.9.4
pyglet==1.3.2
Pygments==2.5.2
pyparsing==2.4.2
pypiwin32==223
pytesseract==0.3.0
python-dateutil==2.8.0
python-docx==0.8.10
pytz==2019.2
pywin32==227
PyYAML==5.1.2
requests==2.22.0
s3transfer==0.2.1
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
sklearn==0.0
smart-open==1.8.4
sortedcontainers==2.1.0
soupsieve==1.9.5
spacy==2.2.1
srsly==0.1.0
tb-nightly==1.14.0a20190603
tempora==2.1.0
tensorflow==2.0.0b1
tensorflow-hub==0.6.0
termcolor==1.1.0
textblob==0.15.3
tf-estimator-nightly==1.14.0.dev2019060501
thinc==7.1.1
tqdm==4.36.1
urllib3==1.25.6
virtualenv==16.7.5
virtualenv-clone==0.5.3
wasabi==0.2.2
Werkzeug==0.16.0
wincertstore==0.2
wrapt==1.11.2
zc.lockfile==2.0
